# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API de Gerenciamento de Usuários e Peixes",
      "description": "API para gerenciar informações de usuários e peixes.",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "https://api-peixes-cxxg.vercel.app",
        "description": "Servidor de Produção"
      }
    ],
    "paths": {
      "/api/users": {
        "post": {
          "summary": "Cadastrar Usuário",
          "description": "Cria um novo usuário.",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Usuário criado com sucesso."
            },
            "400": {
              "description": "Requisição inválida."
            },
            "500": {
              "description": "Erro interno do servidor."
            }
          }
        },
        "get": {
          "summary": "Listar Usuários",
          "description": "Lista todos os usuários cadastrados.",
          "responses": {
            "200": {
              "description": "Lista de usuários.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Erro interno do servidor."
            }
          }
        }
      },
      "/api/users/{id}": {
        "put": {
          "summary": "Atualizar Usuário",
          "description": "Atualiza um usuário existente.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Usuário atualizado com sucesso."
            },
            "400": {
              "description": "Requisição inválida."
            },
            "404": {
              "description": "Usuário não encontrado."
            },
            "500": {
              "description": "Erro interno do servidor."
            }
          }
        },
        "delete": {
          "summary": "Remover Usuário",
          "description": "Remove um usuário existente.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Usuário removido com sucesso."
            },
            "404": {
              "description": "Usuário não encontrado."
            },
            "500": {
              "description": "Erro interno do servidor."
            }
          }
        }
      },
      "/api/peixes": {
        "post": {
          "summary": "Cadastrar Peixe",
          "description": "Cria um novo peixe.",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Peixe"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Peixe criado com sucesso."
            },
            "400": {
              "description": "Requisição inválida."
            },
            "500": {
              "description": "Erro interno do servidor."
            }
          }
        },
        "get": {
          "summary": "Listar Peixes",
          "description": "Lista todos os peixes cadastrados.",
          "responses": {
            "200": {
              "description": "Lista de peixes.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Peixe"
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Erro interno do servidor."
            }
          }
        }
      },
      "/api/peixes/{id}": {
        "put": {
          "summary": "Atualizar Peixe",
          "description": "Atualiza um peixe existente.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Peixe"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Peixe atualizado com sucesso."
            },
            "400": {
              "description": "Requisição inválida."
            },
            "404": {
              "description": "Peixe não encontrado."
            },
            "500": {
              "description": "Erro interno do servidor."
            }
          }
        },
        "delete": {
          "summary": "Remover Peixe",
          "description": "Remove um peixe existente.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Peixe removido com sucesso."
            },
            "404": {
              "description": "Peixe não encontrado."
            },
            "500": {
              "description": "Erro interno do servidor."
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "User": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          },
          "required": ["name", "email", "password"]
        },
        "Peixe": {
          "type": "object",
          "properties": {
            "ID_usuario": {
              "type": "string"
            },
            "Nome": {
              "type": "string"
            },
            "Especie": {
              "type": "string"
            },
           
            "Alimentacao": {
                "type": "string"
              },
              "Quantidade_comida": {
                "type": "number"
              },
              "Vezes_comida_dia": {
                "type": "number"
              },
              "Imagem_url": {
                "type": "string"
              }
            },
            "required": ["ID_usuario", "Nome", "Especie", "Alimentacao", "Quantidade_comida", "Vezes_comida_dia", "Imagem_url"]
          }
        }
      }
    }
      